修正剩下錯誤
